requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
grpcio-reflection==1.76.0
prometheus-client==0.23.1
aiohttp==3.13.2
apache-skywalking==1.2.0
protobuf==6.33.1
googleapis-common-protos==1.72.0
fastapi==0.115.0
requests==2.32.5
uvicorn==0.40.0
uvicorn-worker==0.4.0
langfuse==3.10.1
makefun==1.16.0
gunicorn==24.1.1
boto3
python-docx==1.1.2
sqlmodel
openai
loguru
docxtpl
langchain==1.0.5
langchain-openai==1.0.2
langchain-core==1.0.4
pymysql
pandas
openpyxl